chore(clean-types): moves error into new errors.py types file
