+++ runem/types_wip/errors.py
@@ -1,0 +1,4 @@
+class FunctionNotFound(ValueError):
+    """Thrown when the test-function cannot be found."""
+
+    pass
